+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+* Fix spock 2.4 compatibility (spock 2.4 moved FEATURE_EXECUTION before initialization, which might cause problems
+  to feature interceptors, trying to access junit store).
+
 ### 1.4.0 (2026-02-04)
 * [Junit 5.12](https://docs.junit.org/5.12.1/release-notes/) compatibility
     - Implement missed extension context API methods